(#3) move application helper classes into the application
This is not really the right thing to do, these should be util classes
but it fixes the immmediate problem without huge refactors
