+++ spec/unit/mcollective/application/shell/prefix_stream_buf_spec.rb
@@ -1,5 +1,5 @@
 require 'spec_helper'
-require 'mcollective/application/shell/prefix_stream_buf'
+require 'mcollective/application/shell'
 
 module MCollective
   class Application
