+++ spec/unit/mcollective/application/shell/watcher_spec.rb
@@ -1,5 +1,5 @@
 require 'spec_helper'
-require 'mcollective/application/shell/watcher'
+require 'mcollective/application/shell'
 
 module MCollective
   class Application
